improvement: display reserve action details in proposal page
